# Доработка VideoObserver