… dist/vot-min.user.js
… dist/vot.user.js