Улучшена производительность

```
===========================
  VideoObserver Benchmark  
===========================

Platform info:
==============
   Linux 6.1.43 x64
   Node.JS: 22.6.0
   V8: 12.4.254.21-node.18
   CPU: AMD EPYC × 4
   Memory: 8 GB

Suite: VideoObserver Comparison
===============================

✔ New         5,074 ops/sec
✔ Old         2,729 ops/sec

   New            0%      (5,074 ops/sec)   (avg: 197μs)
   Old       -46.21%      (2,729 ops/sec)   (avg: 366μs)

┌─────┬────────────────────────────────────────────────────┐
│ New │ ██████████████████████████████████████████████████ │
├─────┼────────────────────────────────────────────────────┤
│ Old │ ███████████████████████████                        │
└─────┴────────────────────────────────────────────────────┘
-----------------------------------------------------------------------
```